Auto select best performing model to use, define which error type to prioritise in function call
The error metric to use to dtermine which model performs best can now be defined in the function call to optimise_train_model(). The options are F1, accuracy or recall. The confiusion matrices for all three options are always plotted. The model that performs best accoridng to the selected metric is automatically used in the image analysis.