Add function to append training dtaa from images
new function allows spectral data to be added from uav images to the training set. Areas identified manually as homogenous and of known label can be used to generate rows for the trainign set. WARNING: pixel coordinates in committed verson are only preliminary, roughly selected areas - this will be updated in next commits