Parallelize Update
Realized error in previous attempts using Dask and joblib: random forest classifier is already running in parallel (by definition - it is multiple trees and a voting classifier) running on all available cores using njobs=-1. Therefore any attempt to enhance the parallelization using other means leads to a performance decrease relative to straightforward setting of njobs. Reducing the number of estimators to 64 now makes the prediction function faster than creating the albedo array, which should be the next priority to optimize.