Initial commit for fast Sentinel-2 classifier with GIMP mask
New code applies classifiier to Seninel 2 data. Instead of taking sub-areas as in past versions, this code masks out non-ice areas using GIMP mask so that the classifier can be applied to the whole S2 tile. Now returns a NetCDF of the relevant dtaa and metadata, consistent with the UAv classifier